refactor: overhaul dynamic string APIs
- Rename yio_aprintf to yio_asprintf (standard alignment).
- Rename yio_reaprintf to yio_append (simplified name).
- Implement storage reuse in yio_asprintf when a non-NULL pointer is passed.
- Remove yio_formatf, yio_reformatf, and related 'reformat' functions/macros.
- Update all tests and examples to use the new dynamic string APIs.
- Remove redundant casts and use size_t for precision in formatting backends.

+++ src/yio/private/yio_float_strfrom_ryu.c
@@ -145,8 +145,8 @@ int YYIO_float_astrfrom_ryu$1_in(YYIO_string *res, int precision0, char spec, YY
 	}
 	// https://github.com/ulfjack/ryu/issues/197
 	const size_t minsize = 2000;
-	const int precision = (int)yio_precision_get_default(precision0, 6);
-	const size_t toalloc = minsize + (size_t)precision + 1;
+	const int precision = yio_precision_get_default(precision0, 6);
+	const size_t toalloc = minsize + precision + 1;
 	int err = YYIO_string_reserve(res, toalloc);
 	if (err) return err;
 	int len = 0;
@@ -159,9 +159,9 @@ int YYIO_float_astrfrom_ryu$1_in(YYIO_string *res, int precision0, char spec, YY
 		len = d2s_buffered_n(val, buf);
 		exp_to_upper(buf, spec == 'G');
 	} else if (spec == 'f' || spec == 'F') {
-		len = d2fixed_buffered_n(val, (uint32_t)precision, buf);
+		len = d2fixed_buffered_n(val, precision, buf);
 	} else if (spec == 'e' || spec == 'E') {
-		len = d2exp_buffered_n(val, (uint32_t)precision, buf);
+		len = d2exp_buffered_n(val, precision, buf);
 		exp_to_upper(buf, spec == 'E');
 	} else {
 		return YYIO_RYU_FALLBACK$1(res, precision0, spec, val);
